:truck: split `test_admin_tenants` into multiple modules, for each endpoint
